Fixes a bug when the IPN/webhook is received with "complete" status.
When the "complete" IPN message was coming through this code would throw a RuntimeException with "Wrong BitPay status..." message, this is because this code was looking for a status "completed" instead of "complete".

Possibly the original developer might have gotten mixed up because Woocommerce order status is named "completed" but the BitPay invoice status is named "complete" which is very similar.